update if, variables, functions and add pointers
