--- ch1/conditional-statements/variables/main.go
@@ -1,12 +1,0 @@
-package main
-
-func main() {
-	var myAge int
-	myAge = 27
-
-	println("My age is: ", myAge)
-
-	yourAge := 28
-
-	println("Your age is: ", yourAge)
-}
