+++ ch1/pointers/main.go
@@ -1,0 +1,11 @@
+package main
+
+func main() {
+	i := 1337
+
+	pointer := &i
+	println(*pointer)
+
+	*pointer = 23
+	println(i)
+}
